+++ resources/ReleaseNotes.md
@@ -7,6 +7,7 @@
 
 # Software
 
+- Add support for shaded regions in `Stackedplot` for events with durations/end times
 - Add `filter` method for science data
 - Add option to not save figure as ".fig"
 - Fix issue with loading data from different time zones